move visitors in visitors/ directory
